Доработал задание 4.1

+++ Yrok4/dz4.1.py
@@ -1,4 +1,4 @@
 print("Введите стороны прямоугольника")
-x, y = map(int, input().split())
+x, y = map(float, input().split())
 print("Площадь равна:", x * y)
 print("Периметр равен:", (x + y) * 2)
